requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
eval_type_backport==0.2.2
exceptiongroup==1.3.0
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
magic-filter==1.0.12
mistralai==1.7.0
multidict==6.4.3
packaging==25.0
playwright==1.52.0
pluggy==1.6.0
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
text-unidecode==1.3
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
